fix: `randomAttendanceCode` 메서드를 service로 이동
dto는 단순히 값을 넘겨주는 역할인데 이외의 기능이 있으면 안된다고 생각하여 service로 이동
